+++ apps/examples/blinkid-preload/CHANGELOG.md
@@ -1,9 +1,19 @@
 # example-blinkid-preload
 
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/camera-manager@7.2.4
+  - @microblink/blinkid-ux-manager@7.4.2
+  - @microblink/blinkid-core@7.4.2
+
 ## 1.1.11
 
 ### Patch Changes
 
+- Fix getting result and retry
 - Updated dependencies
   - @microblink/camera-manager@7.2.3
   - @microblink/blinkid-ux-manager@7.4.1