+++ packages/blinkid/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @microblink/blinkid
 
+## 7.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/camera-manager@7.2.4
+  - @microblink/blinkid-ux-manager@7.4.2
+  - @microblink/blinkid-core@7.4.2
+
 ## 7.4.1
 
 ### Patch Changes